// src/enums/vscodeCommandEnum.ts
export enum VscodeCommandEnum {
    IMPORT_NOVEL = 'novelReader.importNovel',
    OPEN_CHAPTER = 'novelReader.openChapter',
    SHOW_CURRENT_NOVEL = 'novelReader.showCurrentNovel',
    SHOW_CURRENT_CHAPTER = 'novelReader.showCurrentChapter',
    NEXT_CHAPTER = 'novelReader.nextChapter',
    PREVIOUS_CHAPTER = 'novelReader.previousChapter',
    REFRESH_SIDEBAR = 'novelReader.refreshSidebar',
    DELETE_NOVEL = 'novelReader.deleteNovel'
}